Refactor BaseMutationWithMetadata class
